feat: extract models cli group
