+++ application/backend/src/cli.py
@@ -1,12 +1,10 @@
 """Command line interface for interacting with the Physical AI Studio application."""
 
-import asyncio
 import sys
-from pathlib import Path
-from uuid import UUID
 
 import click
 
+from cli.models import models
 from db.engine import get_sync_db_session
 from db.migration import MigrationManager
 from db.schema import (
@@ -122,64 +120,7 @@ def migrate() -> None:
         sys.exit(1)
 
 
-# =============================================================================
-# Models subcommand group
-# =============================================================================
-
-
-@cli.group()
-def models() -> None:
-    """Model management commands."""
-
-
-@models.command("import-dir")
-@click.option("--source-dir", required=True, type=click.Path(exists=True, file_okay=False, path_type=Path))
-@click.option("--project-id", required=True, type=click.UUID)
-@click.option("--dataset-id", required=True, type=click.UUID)
-@click.option("--model-name", required=True, type=str)
-@click.option("--move/--copy", default=False, show_default=True)
-@click.option("--base-model-id", type=click.UUID, default=None)
-@click.option("--version", type=int, default=1, show_default=True)
-def import_dir(
-    source_dir: Path,
-    project_id: UUID,
-    dataset_id: UUID,
-    model_name: str,
-    move: bool,
-    base_model_id: UUID | None,
-    version: int,
-) -> None:
-    """Import a model from an existing folder (copy or move) trained by Physical AI Studio.
-
-    The dataset_id should reference a dataset that uses the same environment as the
-    original training. The inference UI uses this dataset to determine which environment
-    to load when running the model.
-    """
-    from services.model_import_service import ModelImportService
-
-    click.echo(f"Importing model from folder: {source_dir}")
-    click.echo(f"Mode: {'move' if move else 'copy'}")
-
-    async def _run_import() -> None:
-        service = ModelImportService()
-        model = await service.import_model_directory(
-            source_dir=source_dir,
-            project_id=project_id,
-            dataset_id=dataset_id,
-            model_name=model_name,
-            move=move,
-            base_model_id=base_model_id,
-            version=version,
-        )
-        click.echo("Model imported successfully!")
-        click.echo(f"Model ID: {model.id}")
-        click.echo(f"Model path: {model.path}")
-
-    try:
-        asyncio.run(_run_import())
-    except Exception as e:
-        click.echo(f"Model import failed: {e}")
-        sys.exit(1)
+cli.add_command(models)
 
 
 if __name__ == "__main__":
